Realizada a implementação da autenticação OAuth2 e políticas de acessos ao model de Client

+++ app/Http/Controllers/ClientsController.php
@@ -4,33 +4,47 @@ namespace App\Http\Controllers;
 
 use App\Client;
 use App\Http\Requests\ClientsRequest as Request;
+use Carbon\Carbon;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Cache;
 
 class ClientsController extends Controller
 {
 
     public function index()
     {
-        return Client::all();
+        $clients = Cache::remember('api::clients', Carbon::now()->addMinutes(10), function () {
+            return Client::all();
+        });
+        return $clients;
     }
 
     public function store(Request $request)
     {
-        return Client::create($request->all());
+        Cache::forget('api::clients');
+        $data = $request->all();
+        $data['user_id'] = Auth::user()->id;
+        return Client::create($data);
     }
 
     public function show(Client $client)
     {
+        $this->authorize('update', $client);,
         return $client;
     }
 
     public function update(Request $request, Client $client)
     {
+        $this->authorize('update', $client);
+        Cache::forget('api::clients');
         $client->update($request->all());
         return $client;
     }
 
     public function destroy(Client $client)
     {
+        $this->authorize('delete', $client);
+        Cache::forget('api::clients');
         $client->delete();
         return $client;
     }
